welcome to 64bit

+++ include/dstd/cstdint.hpp
@@ -2,9 +2,9 @@
 
 using int8_t   = signed char;
 using int16_t  = signed short;
-using int32_t  = signed int;
+using int32_t  = signed long;
 using int64_t  = signed long long;
 using uint8_t  = unsigned char;
 using uint16_t = unsigned short;
-using uint32_t = unsigned int;
+using uint32_t = unsigned long;
 using uint64_t = unsigned long long;
